Fix Windows build

+++ xenium/nikolaev_queue.hpp
@@ -114,7 +114,7 @@ private:
     }
 
     ~node() override {
-      std::size_t eidx;
+      std::uint64_t eidx;
       while (_allocated_queue.dequeue<false, pop_retries>(eidx, entries_per_node, remap_shift)) {
         reinterpret_cast<T&>(_storage[eidx]).~T();
       }
@@ -127,7 +127,7 @@ private:
     }
 
     bool try_push(value_type&& value) {
-      std::size_t eidx;
+      std::uint64_t eidx;
       if (!_free_queue.dequeue<false, pop_retries>(eidx, entries_per_node, remap_shift)) {
         _allocated_queue.finalize();
         return false;
@@ -149,7 +149,7 @@ private:
     }
 
     bool try_pop(value_type& result) {
-      std::size_t eidx;
+      std::uint64_t eidx;
       if (!_allocated_queue.dequeue<false, pop_retries>(eidx, entries_per_node, remap_shift)) {
         return false;
       }
